SeekBar: fix progressbar glitch with custom min-max values
when SeekBar is called, it does:
mSeekBar.setMax(mMax - mMin);
assuming mMin=0 and mMax=100 if they are not set with settings:min or max
through the related menu xml.
So if we set one of those values in the related menu java class with the
available public voids (setMax(value) and setMin(value)) the seekbar doesn't
get the new min and max thus the user can't slide it to the max position.

To reply:
- set settings:max (or android:max according to your seekbar implementation)
to the wanted value, e.g. 255;
- set the min value through your menu java class with SeekBarPreference.setMin(value);
- try to move the progressbar thumb to the max position, it won't reach it.

Change-Id: I24161b7390a577e2050825ec6276c59ca917f9a9